CORE-1366: share IsConfigExtension via k8sutils and load actions in instrumentor
Move manifest-based config-extension detection into k8sutils so instrumentor
and autoscaler use the same check, and load the actions catalog at instrumentor
startup so AddedToSourcesConfig is actually set.
