+++ autoscaler/controllers/common/configextactions.go
@@ -4,13 +4,11 @@ import (
 	"encoding/json"
 	"slices"
 
-	"github.com/odigos-io/odigos/actions"
-	actionsv1 "github.com/odigos-io/odigos/api/actions/v1alpha1"
 	"github.com/odigos-io/odigos/api/k8sconsts"
 	odigosv1 "github.com/odigos-io/odigos/api/odigos/v1alpha1"
-	odigosactions "github.com/odigos-io/odigos/api/odigos/v1alpha1/actions"
 	odigoscommon "github.com/odigos-io/odigos/common"
 	"github.com/odigos-io/odigos/common/consts"
+	actionutil "github.com/odigos-io/odigos/k8sutils/pkg/action"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 )
@@ -22,12 +20,6 @@ func ConvertActionsToConfigExtensionProcessors(actions odigosv1.ActionList) []od
 		processors = append(processors, actionToConfigExtensionProcessor(processorType, signals))
 	}
 	return processors
-}
-
-// IsConfigExtension reports whether the action type uses the
-// odigosConfigExtension config mechanism, regardless of whether it is disabled.
-func IsConfigExtension(action *odigosv1.Action) bool {
-	return len(configExtensionProcessorTypes(action)) > 0
 }
 
 // IsLegacyConfigExtensionProcessorType reports Processor CR types that used to back
@@ -69,22 +61,7 @@ func extractConfigExtActions(action *odigosv1.Action) []string {
 	if action.Spec.Disabled {
 		return []string{}
 	}
-	return configExtensionProcessorTypes(action)
-}
-
-func configExtensionProcessorTypes(action *odigosv1.Action) []string {
-	manifest, ok := actionManifestForCRD(action)
-	if !ok {
-		return []string{}
-	}
-
-	processors := []string{}
-	for _, processorMetadata := range manifest.Spec.Processors {
-		if processorMetadata.ConfigMechanism == "odigosConfigExtension" {
-			processors = append(processors, processorMetadata.Type)
-		}
-	}
-	return processors
+	return actionutil.ConfigExtensionProcessorTypes(action)
 }
 
 func actionToConfigExtensionProcessor(processorType string, signals []odigoscommon.ObservabilitySignal) odigosv1.Processor {
@@ -108,38 +85,5 @@ func actionToConfigExtensionProcessor(processorType string, signals []odigoscomm
 			CollectorRoles:  []odigosv1.CollectorsGroupRole{odigosv1.CollectorsGroupRoleClusterGateway},
 			ProcessorConfig: runtime.RawExtension{Raw: configJSON},
 		},
-	}
-}
-
-// actionManifestForCRD returns the static YAML catalog entry for the Action CRD's type.
-func actionManifestForCRD(action *odigosv1.Action) (actions.Action, bool) {
-	return actions.GetActionByType(actionCatalogType(action))
-}
-
-// actionCatalogType returns the catalog metadata.type for the configured Action config.
-func actionCatalogType(action *odigosv1.Action) string {
-	switch {
-	case action.Spec.AddClusterInfo != nil:
-		return actionsv1.ActionNameAddClusterInfo
-	case action.Spec.DeleteAttribute != nil:
-		return actionsv1.ActionNameDeleteAttribute
-	case action.Spec.RenameAttribute != nil:
-		return actionsv1.ActionNameRenameAttribute
-	case action.Spec.PiiMasking != nil:
-		return odigosactions.ActionNamePiiMasking
-	case action.Spec.K8sAttributes != nil:
-		return "K8sAttributesResolver"
-	case action.Spec.URLTemplatization != nil:
-		return odigosactions.ActionNameURLTemplatization
-	case action.Spec.SpanRenamer != nil:
-		return odigosactions.ActionSpanRenamer
-	case action.Spec.ExtractAttribute != nil:
-		return odigosactions.ActionNameExtractAttribute
-	case action.Spec.DbQueryTemplatization != nil:
-		return odigosactions.ActionNameDbQueryTemplatization
-	case action.Spec.InferDbAttributes != nil:
-		return odigosactions.ActionNameInferDbAttributes
-	default:
-		return ""
 	}
 }
