+++ instrumentor/controllers/agentenabled/action_controller.go
@@ -6,12 +6,13 @@ import (
 	odigosv1 "github.com/odigos-io/odigos/api/odigos/v1alpha1"
 	"github.com/odigos-io/odigos/distros"
 	"github.com/odigos-io/odigos/instrumentor/controllers/agentenabled/rollout"
+	actionutil "github.com/odigos-io/odigos/k8sutils/pkg/action"
 	odgiosK8s "github.com/odigos-io/odigos/k8sutils/pkg/conditions"
 	"github.com/odigos-io/odigos/k8sutils/pkg/utils"
 	"github.com/odigos-io/odigos/status"
 	addedToSourcesConfig "github.com/odigos-io/odigos/status/action/generated"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
-	"k8s.io/apimachinery/pkg/types"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
@@ -23,31 +24,37 @@ type ActionReconciler struct {
 }
 
 func (r *ActionReconciler) Reconcile(ctx context.Context, req ctrl.Request) (ctrl.Result, error) {
-	// This reconciler is fired everytime an agent-level action (URLTemplatization,
-	// SpanRenamer, or config-extension actions that update sources config) is
-	// created, updated or deleted. Reconcile all workloads so InstrumentationConfigs
-	// pick up the change, then sync AddedToSourcesConfig for config-ext actions.
+	// Load the Action before reconcileAll so status is updated from the same
+	// object generation that drove source config reconciliation.
+	action := &odigosv1.Action{}
+	err := r.Get(ctx, req.NamespacedName, action)
+	if err != nil {
+		if !apierrors.IsNotFound(err) {
+			return ctrl.Result{}, err
+		}
+		// Action deleted: still reconcile sources so its config is removed.
+		action = nil
+	}
+
 	result, err := reconcileAll(ctx, r.Client, r.DistrosProvider, r.RolloutConcurrencyLimiter)
 	if err != nil {
 		return utils.K8SUpdateErrorHandler(err)
 	}
-	if !result.IsZero() {
-		return result, nil
+	// Sync status even when reconcileAll asks to requeue (e.g. rollout wait):
+	// InstrumentationConfigs are already updated before rollout requeue.
+	if action != nil {
+		if syncErr := syncActionAddedToSourcesConfig(ctx, r.Client, action); syncErr != nil {
+			return utils.K8SUpdateErrorHandler(syncErr)
+		}
 	}
-	return utils.K8SUpdateErrorHandler(syncActionAddedToSourcesConfig(ctx, r.Client, req.NamespacedName))
+	return result, nil
 }
 
 // syncActionAddedToSourcesConfig sets AddedToSourcesConfig when this action is a
 // config-extension action that drives source InstrumentationConfig (added or
 // removed-because-disabled), otherwise clears a stale condition if present.
-func syncActionAddedToSourcesConfig(ctx context.Context, c client.Client, key types.NamespacedName) error {
-	action := &odigosv1.Action{}
-	err := c.Get(ctx, key, action)
-	if err != nil {
-		return err
-	}
-
-	if !isConfigExtensionSourcesAction(action) {
+func syncActionAddedToSourcesConfig(ctx context.Context, c client.Client, action *odigosv1.Action) error {
+	if !actionutil.IsConfigExtension(action) {
 		return clearAddedToSourcesConfig(ctx, c, action)
 	}
 
@@ -55,12 +62,6 @@ func syncActionAddedToSourcesConfig(ctx context.Context, c client.Client, key ty
 		return setAddedToSourcesConfigReason(ctx, c, action, addedToSourcesConfig.AddedToSourcesConfigConfigRemovedDisabled)
 	}
 	return setAddedToSourcesConfigReason(ctx, c, action, addedToSourcesConfig.AddedToSourcesConfigConfigUpdated)
-}
-
-func isConfigExtensionSourcesAction(action *odigosv1.Action) bool {
-	return action.Spec.PiiMasking != nil ||
-		action.Spec.DbQueryTemplatization != nil ||
-		action.Spec.InferDbAttributes != nil
 }
 
 func setAddedToSourcesConfigReason(ctx context.Context, c client.Client, action *odigosv1.Action, reason status.Reason) error {
